requirements.txt
aioredis==1.0.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.4
channels==2.1.2
channels-redis==2.1.1
constantly==15.1.0
cryptography==3.2.1
daphne==2.5.0
Django==2.2.2
djangorestframework==3.12.2
gunicorn==20.1.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==0.5.6
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
pytz==2020.1
redis==3.5.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
whitenoise==5.3.0
zope.interface==5.2.0
